Attendance/Information
Created repository, and added team attendance.
Get familliar with the sheet. If there is a 1 next to your name on
attendance just click the comment and say why. This isn't a trouble
thing. It's so we can keep track and just know what's going on.

Also, the team information: add your phone number if it is not there,
or update if it is wrong. Add what else you have going on as well and
your major!